split site model into concerns
